requirements/base.txt
bcrypt==3.1.2
boto3==1.4.4
cryptography==1.7.1
Django<1.11
django-filter==1.0.1
django-rest-swagger==2.1.1
djangorestframework==3.5.3
docker-py==1.10.6
drf-nested-routers==0.11.1
GitPython==2.1.1
hashids==1.2.0
hiredis==0.2.0
humanize==0.5.1
nameparser==0.5.1
passlib==1.7.0
psycopg2==2.6.2
Pygments==2.2.0
python-dateutil==2.6.0
python-magic==0.4.12
pytz==2016.10
rauth==0.7.3
dj-database-url==0.4.2
bcrypt==3.1.2
drf-ujson==1.2
ujson==1.35
django-redis==4.7.0
django-extensions==1.7.5
gunicorn==19.6.0